+++ sitewit/__init__.py
@@ -1,4 +1,4 @@
 """Python client for the SiteWit API."""
 
-__version__ = '0.3.2'
+__version__ = '0.4.0'
 __url__ = 'https://github.com/yola/sitewit'