+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # [Changelog](https://github.com/yola/sitewit/releases)
 
+## 0.4.0
+
+* Add service method to change account owner
+* Add 2 model methods to associate account with either new user or existing
+  user
+* Fix `Account` model initialization
+* Change `Account.create` to accept `site_id` instead of Site instance
+* Fix `Account.create` to populate requested attributes properly.
+
 ## 0.3.2
 
 * Add `billing_date` attribute to `Subscription`